used switch cases and formatted the code to 2 tab spaces used switch cases and formatted t
