--- src/main/java/com/demo/Model/PlanType.java
@@ -1,16 +1,0 @@
-package com.demo.Model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-public enum PlanType {
-    SIM_ONLY_PLANS("simo"),
-    SIM_AND_PHONE_PLANS("phones");
-
-
-    @Getter
-    private final String planType;
-
-
-}
