--- src/main/java/com/demo/Model/Price.java
@@ -1,16 +1,0 @@
-package com.demo.Model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Price {
-
-    private Double net;
-    private Double totalPrice;
-    private Double totalGst;
-    private Double cGst;
-    private Double sGst;
-
-}
