+++ src/main/java/com/demo/configuration/SwaggerConfiguration.java
@@ -14,7 +14,7 @@ import java.util.Collections;
 
 @Configuration
 @EnableSwagger2
-public class SwaggerConfig {
+public class SwaggerConfiguration {
 
 
     //Docker bean created for Swagger
