+++ src/main/java/com/demo/exceptions/ErrorResponse.java
@@ -1,16 +1,12 @@
 package com.demo.exceptions;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
 public class ErrorResponse {
-
-private String message;
-private String errorCode;
-private String errorDescription;
-
-
+  private String message;
+  private String errorCode;
+  private String errorDescription;
 }
