+++ src/main/java/com/demo/exceptions/PlanExceptions.java
@@ -1,11 +1,9 @@
 package com.demo.exceptions;
 
-
 public class PlanExceptions extends RuntimeException {
 
-    public PlanExceptions(String message){
-        super(message);
-    }
-
+  public PlanExceptions(String message) {
+    super(message);
+  }
 
 }
