+++ src/main/java/com/demo/model/Plans.java
@@ -1,4 +1,4 @@
-package com.demo.Model;
+package com.demo.model;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
