+++ src/main/java/com/demo/model/Price.java
@@ -1,0 +1,14 @@
+package com.demo.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Price {
+  private Double net;
+  private Double totalPrice;
+  private Double totalGst;
+  private Double cGst;
+  private Double sGst;
+}
