+++ src/main/java/com/demo/service/PlansService.java
@@ -1,10 +1,10 @@
 package com.demo.service;
 
-import com.demo.Model.PlanType;
-import com.demo.Model.Plans;
 import com.demo.dao.PlansDao;
 import com.demo.exceptions.ErrorMessage;
 import com.demo.exceptions.PlanExceptions;
+import com.demo.model.PlanType;
+import com.demo.model.Plans;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,23 +14,15 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PlansService {
 
-    private final PlansDao plansDao;
+  private final PlansDao plansDao;
 
-
-    public List<Plans> getPlans(String planType) {
-
-        if (!(planType.equalsIgnoreCase(PlanType.SIM_ONLY_PLANS.getPlanType())
-                || planType.equalsIgnoreCase(PlanType.SIM_AND_PHONE_PLANS.getPlanType())))
-        {
-            throw new PlanExceptions(ErrorMessage.WRONG_PLANTYPE.getMessage());
-        }
-
-        if (PlanType.SIM_ONLY_PLANS.getPlanType().equalsIgnoreCase(planType))
-        {
-            return plansDao.getPlans();
-        } else
-            return null;
-
-
+  public List<Plans> getPlans(PlanType planType) {
+    switch (planType) {
+      case SIM_ONLY_PLANS:
+        return plansDao.getPlans();
+      case SIM_AND_PHONE_PLANS:
+      default:
+        throw new PlanExceptions(ErrorMessage.WRONG_PLANTYPE.getMessage());
     }
+  }
 }
